test(flydsl): move the runtime plumbing tests out of the rmsnorm file
Five tests in test_rmsnorm_flydsl.py never touched rmsnorm: two cover
validate_arch, one covers empty_placeholder, and two assert that a ROCm-only
install fails to reach for CuTe (quack/__init__.py and the pytest plugin).
They live in tests/test_flydsl_runtime.py now.

Both modules need the ROCm/gfx950/flydsl gate at import time -- for pytestmark
and to guard the quack imports -- so it moves to tests/flydsl_env.py rather
than being duplicated or turned into a fixture.
