+++ tests/test_rmsnorm_flydsl.py
@@ -1,34 +1,13 @@
 # Copyright (c) 2026, Tri Dao.
 
-import dataclasses
-import importlib.util
 import logging
-import re
-import sys
 
 import pytest
 import torch
+from flydsl_env import CAN_RUN as _CAN_RUN
+from flydsl_env import DEVICE as _DEVICE
+from flydsl_env import SKIP_REASON as _SKIP_REASON
 
-_HAS_ROCM_GPU = torch.version.hip is not None and torch.cuda.is_available()
-_DEVICE = (
-    torch.device("cuda", torch.cuda.current_device()) if _HAS_ROCM_GPU else torch.device("cuda")
-)
-_ARCH = (
-    torch.cuda.get_device_properties(_DEVICE).gcnArchName.split(":", 1)[0]
-    if _HAS_ROCM_GPU
-    else None
-)
-_HAS_FLYDSL = importlib.util.find_spec("flydsl") is not None
-_CAN_RUN = _HAS_ROCM_GPU and _ARCH == "gfx950" and _HAS_FLYDSL
-
-if not _HAS_ROCM_GPU:
-    _SKIP_REASON = "requires a ROCm GPU"
-elif _ARCH != "gfx950":
-    _SKIP_REASON = "FlyDSL RMSNorm currently requires gfx950"
-elif not _HAS_FLYDSL:
-    _SKIP_REASON = "requires flydsl"
-else:
-    _SKIP_REASON = ""
 pytestmark = pytest.mark.skipif(not _CAN_RUN, reason=_SKIP_REASON)
 
 if _CAN_RUN:
@@ -449,38 +428,6 @@ def test_misaligned_contiguous_storage_offsets_force_allocation():
     _assert_close(out, expected, x.dtype)
 
 
-@pytest.mark.parametrize("arch", ("9.5.0", "gfx942"), ids=("dotted", "wrong-gfx"))
-def test_compile_target_that_the_device_never_reports_is_rejected(monkeypatch, arch):
-    # ARCH reaches FlyDSL's ROCDL pipeline as `chip` verbatim, so an override
-    # the device never reports has to fail here. Normalizing "9.5.0" to gfx950
-    # would let it past this check and fail later inside the compiler instead.
-    # get_backend() memoizes one instance per (name, arch), so patching the
-    # instance is what validate_arch reads back.
-    backend = fly_runtime.flyc.get_backend()
-    monkeypatch.setattr(backend, "target", dataclasses.replace(backend.target, arch=arch))
-
-    with pytest.raises(ValueError, match=f"FlyDSL compiles for {re.escape(arch)}"):
-        fly_runtime.validate_arch(_DEVICE, frozenset({_ARCH}), "RMSNorm")
-
-
-def test_every_caller_gets_its_own_supported_set_checked():
-    # A per-device arch cache that also gated the supported check would let the
-    # second kernel inherit the first one's verdict on this GPU.
-    assert fly_runtime.validate_arch(_DEVICE, frozenset({_ARCH}), "RMSNorm") == _ARCH
-
-    with pytest.raises(ValueError, match=f"supports gfx900; cuda:\\d+ is {_ARCH}"):
-        fly_runtime.validate_arch(_DEVICE, frozenset({"gfx900"}), "Elsewhere")
-
-
-def test_placeholders_are_shared_across_spellings_of_one_device():
-    # Callers hand over whatever x.device gave them, so an unindexed "cuda" and
-    # an explicit "cuda:N" have to land on the same cached tensor.
-    bare = fly_runtime.empty_placeholder(torch.device("cuda"), torch.float32)
-    indexed = fly_runtime.empty_placeholder(_DEVICE, torch.float32)
-    assert bare is indexed
-    assert bare.numel() == 0 and bare.device.index == _DEVICE.index
-
-
 def test_empty_rows_return_without_compiling(monkeypatch):
     x = torch.empty(0, 192, device=_DEVICE, dtype=torch.float16)
     weight = torch.ones(192, device=_DEVICE, dtype=torch.float32)
@@ -501,29 +448,6 @@ def test_empty_rows_return_without_compiling(monkeypatch):
     assert residual_out.shape == x.shape and residual_out.dtype == torch.float32
     assert residual_out is not x
     assert rstd.shape == x.shape[:-1] and rstd.dtype == torch.float32
-
-
-def test_rocm_cute_exports_fail_without_submodule_fallback():
-    with pytest.raises(ImportError, match="requires the CUDA/CuTe backend"):
-        __import__("quack", fromlist=("rmsnorm",))
-
-    assert "quack.rmsnorm" not in sys.modules
-    assert not any(name == "cutlass" or name.startswith("cutlass.") for name in sys.modules)
-
-
-def test_async_compile_is_rejected_before_loading_cute():
-    from quack.testing import pytest_plugin
-
-    class AsyncCompileConfig:
-        @staticmethod
-        def getoption(*_args, **_kwargs):
-            return 1
-
-    with pytest.raises(pytest.UsageError, match="unavailable on ROCm"):
-        pytest_plugin.pytest_configure(AsyncCompileConfig())
-
-    assert "quack.cache.jit" not in sys.modules
-    assert not any(name == "cutlass" or name.startswith("cutlass.") for name in sys.modules)
 
 
 def test_invalid_inputs():
